GPU: Shader Tool: Optimize Tokenizer and Parser
This makes the following changes:
- Break tokenizer into 2 passes (tokenize, then merge)
- Make tokenizer >1GB/s fast
- Made token type lookup almost as fast
- Limit the input to 4GB (32bit string ranges)
- Use one pass fill for token to scope indirection table
- use always inline for small functions
- use std vector instead of std stack
- Prealocate vectors to avoid checking for size
- Add string_view interface for tokens

Pull Request: https://projects.blender.org/blender/blender/pulls/152437

+++ source/blender/gpu/shader_tool/utils.hh
@@ -28,7 +28,7 @@ struct IndexRange {
   int64_t start;
   int64_t size;
 
-  IndexRange(size_t start, size_t size) : start(start), size(size) {}
+  IndexRange(int64_t start, int64_t size) : start(start), size(size) {}
 
   bool overlaps(IndexRange other) const
   {
@@ -47,11 +47,11 @@ struct IndexRange {
 
 /** Poor man's OffsetIndices. */
 struct OffsetIndices {
-  std::vector<size_t> offsets;
+  std::vector<uint32_t> offsets;
 
   IndexRange operator[](const int64_t index) const
   {
-    return {offsets[index], offsets[index + 1] - offsets[index]};
+    return {int64_t(offsets[index]), int64_t(offsets[index + 1] - offsets[index])};
   }
 
   void clear()
